requirements.txt
langchain==0.2.0
langchain-groq==0.1.3
langchain-community==0.2.0
langgraph==0.0.65
faiss-cpu==1.8.0
sentence-transformers==2.7.0
rank-bm25==0.2.2
streamlit==1.35.0
python-dotenv==1.0.1
pypdf==4.2.0
